Step 9: Introduce local variables consist of a character
This patch introduces the local variables consist of a character, i.e.,
a, b, c, ..., z, to which the numbers can be assigned. The assigned
variables can be used in the arithmetic operations.

To allow the variable assignments and their uses, the compiler is
modified to parse multiple statements that are comprised of the single
expressions ending with semicolons. Therefore the following inputs are
all valid.

- a=42;
- a=0; b+1;
- a = 1; b = 2; c = a + b
- a=1; b=2; c=3; d=7; a*b*c*d;

Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>