Step 10: Allow local variables to be multiple characters
Local characters are restricted to be comprised of the single characters
in the previous step. This patch modifies them to accept multiple
characters as their names.

This patch contains the refactoring of the way to hold the parsed AST
nodes as well . The input is considered a sequence of statements with
local variables, which is embraced by a function. In other words the
compiler parses the input as the single function and generates the
assembly code corresponds to the function as the content of the main
section.

Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>