Step 11: Add return statement
This patch adds return statement that returns from the current function
setting RAX to the given value or the result of the given expression.
The stack pointer is made point to the base register and the popped
address previously pointed by the base register is loaded to base
register to restore the previous state before the function was called.

Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>