Step 15: Support function definitions
This patch adds the function definition support and now we can define
funcitons that have arguments, which are up to six. More than six
arguments and variable arguments are not supported at the moment.
Arguments are regarded the same as the local variables and they are
pushed to the stack.

main is specified as the entrypoint of the program and every test case
MUST contain the main function or the program cannot be linked
otherwise.

Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>