Step 15: Support function definitions
This patch adds the function definition support and now we can define
funcitons that have arguments, which number is up to six. More than
seven arguments and variable arguments are not supported at the moment.
Arguments are regarded the same as the local variables and they are
pushed to the stack before the statements in the funcion bodies are
executed.

"main" is specified as the entrypoint of the program and every test case
MUST contain the main function or the program cannot run otherwise.

Signed-off-by: Taku Fukushima <f.tac.mac@gmail.com>